Fix for the semicolon multiline STAR format parsing.
This is an upstream code push from the bmrblib project (http://gna.org/projects/bmrblib/) which incorporates pystarlib.  Bmrblib was originally part of relax (http://nmr-relax.com).  See https://mail.gna.org/public/bmrblib-commits/2011-03/msg00001.html for the original change.

This was failing if there was text after the terminating ';', but this is allowed in the STAR format and is present in many BMRB entries!